Get fork() working on Windows
This is done without using Microsoft's internal APIs. MAP_PRIVATE
mappings are copied to the subprocess via a pipe, since Microsoft
doesn't want us to have proper COW pages. MAP_SHARED mappings are
remapped without needing to do any copying. Global variables need
copying along with the stack and the whole heap of anonymous mem.
This actually improves the reliability of the redbean http server
although one shouldn't expect 10k+ connections on a home computer
that isn't running software built to serve like Linux or FreeBSD.

+++ libc/nt/struct/msg.h
@@ -6,8 +6,8 @@
 struct NtMsg {
   int64_t hwnd;
   uint32_t message;
-  uintptr_t wParam;
-  intptr_t lParam;
+  uint64_t wParam;
+  int64_t lParam;
   uint32_t time;
   struct NtPoint pt;
 };
